+++ aockt-test/src/test/kotlin/TestConfig.kt
@@ -17,7 +17,4 @@ object TestConfig : AbstractProjectConfig() {
 
     override val specExecutionMode = SpecExecutionMode.LimitedConcurrency(Runtime.getRuntime().availableProcessors())
     override val specExecutionOrder = SpecExecutionOrder.Lexicographic
-
-    // https://kotest.io/docs/framework/test_output.html
-    override var displayFullTestPath: Boolean? = System.getenv("CI").toBoolean()
 }